requirements.txt
pandas==2.1.3
matplotlib==3.8.2
numpy==1.26.2
scipy==1.11.4
python-dotenv~=1.0.0
pandera~=0.17.2
openpyxl~=3.1.2
et-xmlfile~=1.1.0
pytest~=7.4.3